handled one elemnt linked lists

+++ doubly_linked_lists/5-get_dnodeint.c
@@ -10,13 +10,13 @@
  */
 dlistint_t *get_dnodeint_at_index(dlistint_t *head, unsigned int index)
 {
-	unsigned int i;
+	int i;
 	dlistint_t *it = head;
 
 	if (head == NULL)
 		return (NULL);
 
-	for (i = 0; i <= index - 1; i++)
+	for (i = 0; i <= (int) index - 1; i++)
 	{
 		if (it == NULL)
 			return (NULL);
